% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDirectionality.R
\name{addDirectionality}
\alias{addDirectionality}
\title{Add directionality of interactions}
\usage{
addDirectionality(network_list, sce_list, n_cores = 1L, assay = "logcounts")
}
\arguments{
\item{network_list}{A named list of \code{\link{igraph}} objects containing the networks per clone and potentially the consensus network. The list element containing the consensus network should have the name 'consensus'.}

\item{sce_list}{A named list of \code{\link{SingleCellExperiment-class}} objectd containing the expression data (raw counts, normaized counts and metadata) for all network genes per clone.}

\item{n_cores}{Integer, the number of cores (default: 1).}

\item{assay}{Character, the name of the assay in the SingleCellExperiment objects that should be used for the calculation of gene-gene correlations (default: "logcounts").}
}
\value{
A named list of \code{\link{igraph}} objects after adding the information about the direction of interactions. In addition to the edge attributes in the input \code{network_list}, it contains 3 new attributes in each network:
\describe{
\item{rho}{Numeric, the approximate Spearman's correlation coefficient of the 2 genes' expression profiles that form the edge.}
\item{p.adj}{Numeric, BH-corrected approximate p-value of rho.}
\item{direction}{Character, the direction of the interaction between the 2 genes that form the edge ("+" or "-").}
}
}
\description{
Determines the directionality of each network connection based on a modified Spearman's correlation between the 2 genes' expression profiles (positive expression correlation - activating interaction, negative expression correlation - repressing interaction).
}
\details{
If the networks were inferred using a method that cannot distinguish positive and negative regulatory interactions, it might be useful to add this information for lines of analysis where it makes sense to separate the activated and repressed target genes of a regulator (e.g. for the calculation of eigengenes).

The calculation relies on the approximate version of the Spearman's rho, significance testing and blocking implemented by [scran::correlatePairs]. The results are summarized as 3 new edge attributes in the igraph objects: rho (approximate Spearman's correlation coefficient), p.adj (BH-corrected approximate p-value) and direction (+ or -).
}
\examples{
network_list_scaled_filt_withCons_withDir <- addDirectionality(network_list_scaled_filt_withCons,
                                                               sce_list)
}
