% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free_transformation.R
\name{applyPower}
\alias{applyPower}
\title{Apply soft-thresholding}
\usage{
applyPower(network_list, power, n_cores = 1L)
}
\arguments{
\item{network_list}{A named list of \code{\link{igraph}} objects containing the networks of all replicates.}

\item{power}{Numeric, the power to raise the edge weights to.}

\item{n_cores}{Integer, the number of cores (default: 1).}
}
\value{
A named list of \code{\link{igraph}} objects containing the networks of all replicates after the scale-free transformation on the edge weights
}
\description{
Apply soft-thresholding
}
