% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePresStats.R
\name{calculatePresStats}
\alias{calculatePresStats}
\title{Calculate preservation statistics}
\usage{
calculatePresStats(
  pruned_modules,
  network_list,
  stats = c("cor_adj", "cor_adj_regulator", "cor_kIM"),
  clone2species = NULL,
  jackknife = TRUE,
  n_cores = 1L,
  corr_method = "spearman"
)
}
\arguments{
\item{pruned_modules}{Data frame of the pruned modules with columns 'regulator', 'target' and 'weight' containing the transcriptional regulators, their target genes and the edge weights between each regulator-target pair.}

\item{network_list}{A named list that contains the networks of each clone in an igraph format.}

\item{stats}{Character or character vector specifying which preservation statistics to calculate (one or more of "cor_adj", "cor_adj_regulator", and "cor_kIM").}

\item{clone2species}{A data frame with columns 'clone' and 'species' that specifies which species each clone belongs to. The names of clones should match the names of 'network_list'.}

\item{jackknife}{Logical specifying whether jackknifing should be performed or not (default: TRUE).}

\item{n_cores}{Number of cores.}

\item{corr_method}{The method for the calculation of correlation, one of "spearman" or "pearson".}
}
\value{
Data frame of the preservation statistics.
}
\description{
Calculates one or more preservation statistics - correlation of adjacencies (cor_adj), correlation of regulator-target adjacencies (cor_adj_regulator), and/or correlation of intramodular connectivities (cor_kIM) - for all jackknifed versions of the modules and all clone pairs.
}
\details{
The function calculates one or more preservation statistics adapted from Langfelder et al. 2011 (cor_adj, cor_adj_regulator and cor_kIM) for module and each clone pair. All three statistics quantify how well the module topology is preserved between the networks of two clones. The statistic cor_adj is the correlation of all edge weights within the module in the network of clone1 VS all edge weights within the module in the network of clone2. The statistic cor_adj_regulator is the correlation of all edge weights between the regulator and its module members in the network of clone1 VS all edge weights between the regulator and its module members in the network of clone2. Finally, the statistic cor_kIM is the correlation of the intramodular connectivities per module member gene in the network of clone1 VS the intramodular connectivities per module member gene in the network of clone2.

All statistics assume a joint module assignment (typically derived from the consensus network) but compared topological properties directly between the clonewise networks. In this approach, a module is always defined as the same set of genes (nodes), but the interaction strengths (adjacencies) among these genes could differ from clone to clone; poorly preserved modules are expected to have many, while well-preserved modules are expected to have few such topological differences.

If \code{jackknife} is set to TRUE, the function creates all possible jackknifed versions of each input module by removing each target gene assigned to that module (the regulator is never excluded), then it calculates the preservation statistics for all of these jackknifed module versions in addition to the original module. This way, a confidence interval can be calculated for each module and statistic (see \code{\link{summarizeJackknifeStats}}). Later on in the pipeline, jackknifing can also provide information about which target genes within a conserved/diverged module are particularly responsible for the conservation/divergence (see \code{\link{findConservedDivergedTargets}}). If jackknifing is not desired, please set \code{jackknife} to FALSE. This will also substantially reduce running times.
}
\examples{
pres_stats_jk <- calculatePresStats(pruned_modules, network_list, "cor_kIM", clone2species)
}
