% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTreeStats.R
\name{calculateTreeStats}
\alias{calculateTreeStats}
\title{Calculate tree statistics across all modules}
\usage{
calculateTreeStats(tree_list, n_cores = 1L)
}
\arguments{
\item{tree_list}{A list of trees per module as \code{\link{phylo}} objects with clones as tips. All trees are expected to have a component 'species' that specifies which species each tip belongs to.}

\item{n_cores}{Number of cores.}
}
\value{
A data frame of tree statistics.
}
\description{
Calculates various tree statistics (total tree length, total diversity, diversity within each species, monophyleticity for each species and branch length between each species and all others) for all modules based on a list of trees. The tips of the trees should correspond to clones and the branch lengths of the trees should correspond to dissimilarity of module topology between the clones.
}
\examples{
tree_stats_jk <- calculateTreeStats(trees_jk)
}
