% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clone2species}
\alias{clone2species}
\title{Clone-species conversion}
\format{
A data frame with 9 rows and 2 columns:
\describe{
\item{clone}{Name of the clone/cell line.}
\item{species}{Name of the species.}
}
}
\usage{
clone2species
}
\description{
A data frame that specifies which clone belongs to which species.
}
\keyword{datasets}
