% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consensus_network}
\alias{consensus_network}
\title{Consensus network}
\format{
An [igraph] object with 300 nodes, and has 1 node attribute and 3 edge attributes:

Node attributes:
\describe{
\item{name}{Name of the node (gene).}
}
Edge attributes:
\describe{
\item{weight}{Consensus edge weight/adjacency, the weighted average of clonewise adjacencies.}
\item{n_supporting_clones}{The number of clones where the edge was detected.}
\item{supporting_clones}{The list of clones where the edge was detected.}
\item{rho}{Approximate Spearman's correlation coefficient of the 2 genes' expression profiles that form the edge.}
\item{p.adj}{BH-corrected approximate p-value of rho.}
\item{direction}{Direction of the interaction between the 2 genes that form the edge ("+" or "-").}
}
}
\usage{
consensus_network
}
\description{
Consensus network of the 9 primate clones in the example dataset. For each edge, the consensus adjacency was calculated as the weighted average of clonewise adjacencies using weights that correct for 1) the phylogenetic distances between species and 2) the different numbers of clones per species. If an edge was not detected in certain clone, the adjacency of that clone was regarded as 0 for the calculation of the consensus. The directionality of each edge was determined based on a modified Spearman's correlation between the corresponding 2 genes' expression profiles (positive expression correlation - activating interaction, negative expression correlation - repressing interaction). The correlations were calculated per clone, then the mean correlation was taken across all clones.
}
\keyword{datasets}
