% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertToDT}
\alias{convertToDT}
\title{Convert a list of igraphs to a list of data tables}
\usage{
convertToDT(network_list)
}
\arguments{
\item{network_list}{A list of igraph objects}
}
\value{
A list of data tables with the columns 'from', 'to', 'weight' and any additional edge attributes the input igraphs contain.
}
\description{
Takes care of the conversion between the igraph and data table formats of the networks.
}
