% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertToGraph}
\alias{convertToGraph}
\title{Convert a list of data frames to a list of igraphs}
\usage{
convertToGraph(dt_list, network_genes, n_cores = 1L)
}
\arguments{
\item{dt_list}{A list of data tables that contain the columns 'from', 'to' and 'weight'. Each row should correspond to an edge in the network with 'from' and 'to' as the end nodes and 'weight' as the edge weight. Additional columns will be converted to edge attributes in the igraphs.}

\item{network_genes}{Character vector of all genes in the network.}

\item{n_cores}{Number of cores.}
}
\value{
A list of igraphs.
}
\description{
Takes care of the conversion between data table and igraph formats of the networks.
}
