% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRandomModules.R
\name{createRandomModules}
\alias{createRandomModules}
\title{Create random modules}
\usage{
createRandomModules(pruned_modules, network_genes, seed = 0)
}
\arguments{
\item{pruned_modules}{Data frame of the pruned modules with columns 'regulator', 'target' and 'weight' containing the transcriptional regulators, their target genes and the edge weights between each regulator-target pair.}

\item{network_genes}{Character vector of all genes in the network.
Outputs a random module for each module in 'pruned_modules'. The random modules have the same regulators and contain the same number of target genes as the original modules, but these target genes are randomly drawn from 'network_genes'.}

\item{seed}{Integer, the seed to use for the random sampling.}
}
\value{
Data frame of the random modules.
}
\description{
Creates a matching random module for each actual module.
}
\examples{
random_modules <- createRandomModules(pruned_modules, genes)
}
