% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eigengenes_per_species}
\alias{eigengenes_per_species}
\title{Eigengenes per species}
\format{
A data frame with 6300 rows and 8 columns:
\describe{
\item{cell}{Character, the cell barcode.}
\item{species}{Character, the name of the species.}
\item{pseudotime}{Numeric, inferred pseudotime.}
\item{cell_type}{Character, cell type annotation.}
\item{module}{Character, transcriptional regulator and direction of regulation (in this case always nameOfRegulator(+)).}
\item{eigengene}{Numeric, the eigengene (i.e. the first principal component of the scaled and centered logcounts) of the module.}
\item{mean_expr}{Numeric, the mean of the scaled and centered logcounts across all genes in the module.}
\item{regulator_expr}{Numeric, the scaled and centered logcounts of the regulator.}
}
}
\usage{
eigengenes_per_species
}
\description{
Eigengenes of the pruned modules calculated per species using the activated targets in each module. An eigengene summarizes the expression profile of the module as a whole, mathematically it is the first principal component of the module expression data (i.e. the scaled and centered logcounts subsetted for the activated targets and the regulator of the given module). In this case, the principal component was calculated for each species separately.
}
\keyword{datasets}
