% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterModuleTrees.R
\name{filterModuleTrees}
\alias{filterModuleTrees}
\title{Filter modules based on their tree representations}
\usage{
filterModuleTrees(tree_stats, random_tree_stats, p_adj_cutoff = 0.1)
}
\arguments{
\item{tree_stats}{Data frame of the tree-based statistics for the actual (pruned) modules. Required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{total_tree_length}{Numeric, total tree length per module (typically the median across all jackknife versions of a module.}
\item{within_species_diversity}{Numeric, within-species diveristy per module (typically the median across all jackknife versions of a module.}
}}

\item{random_tree_stats}{Data frame of the tree-based statistics for the random modules. Required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{total_tree_length}{Numeric, total tree length per module (typically the median across all jackknife versions of a module).}
\item{within_species_diversity}{Numeric, within-species diveristy per module (typically the median across all jackknife versions of a module).}
}}

\item{p_adj_cutoff}{Numeric, adjusted p-value cutoff (default: 0.1).}
}
\value{
Data frame of the tree-based statistics for the actual (pruned) modules after filtering.
}
\description{
Filters modules using two tree-based statistics (the total tree length and within-species diversity) by comparing the values of the actual modules to the values of the random modules.
}
\details{
The within-species diversity measures the variability of module connectivity patterns within species and the total tree length measures the variability of module connectivity patterns both within and across species. The within-species diversity is a subset of the total tree, hence we expect a linear relationship between the two statistics. This line explains the detection robustness: modules that have both low within-species diversity and low total tree length are well-preserved both within and across species, meaning that these modules could be robustly detected in all clones, whereas modules, for which both metrics are high, are poorly preserved not just across but also within species, indicating a high detection uncertainty ("wobbliness"). Random modules are expected to fall at the wobbly end of the spectrum, while actual modules are expected to fall at the robust end of the spectrum.

The function removes the modules that are uncertain to detect, i.e. modules the are too similar to the random modules in terms of their total tree length and within-species diversity. This similarity is measured by the probability that the statistic of an actual modules comes from the distribution of the statistics of all random modules (one-sample one-sided z-test with the mean and standard deviation of the random module statistics as the expected mean and standard deviation). If the adjusted p-value of a module is below the specified \code{p_adj_cutoff} for either of the two statistics, i.e. the module is significantly different from the random modules, it is kept.
}
