% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTreeStatsLm.R
\name{fitTreeStatsLm}
\alias{fitTreeStatsLm}
\title{Fit (weighted) linear model between tree-based statistics}
\usage{
fitTreeStatsLm(tree_stats, focus = "overall", weighted_lm = TRUE)
}
\arguments{
\item{tree_stats}{Data frame of the tree-based statistics for the pruned modules.

Columns required in case the focus of interest is conservation and overall divergence:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Integer, the numer of target genes assigned to a regulator.}
\item{total_tree_length}{Numeric, total tree length per module (typically the median across all jackknife versions of the module).}
\item{var_total_tree_length}{Numeric, the variance of total tree lengths across all jackknifed versions of the module (optional, only needed for weighted regression).}
\item{within_species_diversity}{Numeric, within-species diveristy per module (typically the median across all jackknife versions of the module).}
}
Columns required in case the focus of interest is species-specific divergence:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Integer, the numer of target genes assigned to a regulator.}
\item{\{\{species\}\}_to_other_branch_length}{Numeric, the branch length between the species of interest and all others per module (typically the median across all jackknife versions of the module).}
\item{var_\{\{species\}\}_to_other_branch_length}{Numeric, the variance of the branch length between the species of interest and all others across all jackknifed versions of the module (optional, only needed for weighted regression).}
\item{\{\{species\}\}_diversity}{Numeric, within-species diveristy per module (typically the median across all jackknife versions of the module).}
}}

\item{focus}{Character, the focus of interest in terms of cross-species conservation, either "overall" if the focus of interest is conservation and overall divergence, or the name of a species if the focus of interest is the divergence between that particular species and all others.}

\item{weighted_lm}{Logical indicating whether the linear regression should be weighted or not (default: TRUE). If TRUE, \code{tree_stats} is expected to contain the column \code{var_total_tree_length} and the weights will be inversely proportional to these variances. If no jackknifing was performed and thus these variances were not calculated, please set this parameter to FALSE.}
}
\value{
An object of class \code{\link{lm}}.
}
\description{
Fits a (weighted) linear model between the total tree length and within-species diversity (in case the focus of interest is conservation and overall divergence) or between the species-to-other branch length and the diversity of a species (in case the focus of interest is the species-specific divergence).
}
\details{
The linear models output by this function can be used to identify conserved and diverged modules, and to identify target genes within these modules that contribute the most to the conservation/divergence. For details, please see \code{\link{findConservedDivergedModules}} and \code{\link{findConservedDivergedTargets}}.

The focus of interest can be specified using the parameter \code{focus}. If \code{focus} is set to "overall" (default), the linear model will be fit between the total tree length and within-species diversity, and subsequent analysis using \code{\link{findConservedDivergedModules}} and \code{\link{findConservedDivergedTargets}} can identify modules and target genes that are conserved or diverged across all species. If \code{focus} is set to the name of a species in the dataset, the linear model will be fit between the species-to-other branch length and the diversity of that species, and subsequent analysis using \code{\link{findConservedDivergedModules}} and \code{\link{findConservedDivergedTargets}} can identify modules and target genes that are diverged between the species and all others. Please note that if the aim is to find conserved modules, \code{focus} should always be set to "overall".

The function fits the linear model corresponding to the focus of interest by calling \code{\link{lm}}. If a weighted model is desired (default), the weights are defined to be inversely proportional to the variance of the dependent variable (total tree length or the species-to-other branch length). If no jackknifing was performed and thus the variance is unknown, please set \code{weighted_lm} to FALSE.
}
\examples{
lm_overall <- fitTreeStatsLm(tree_stats, focus = "overall")
}
