% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegulators.R
\name{getRegulators}
\alias{getRegulators}
\title{Get transcriptional regulators}
\usage{
getRegulators(sce, source = "jaspar_core", ...)
}
\arguments{
\item{sce}{\code{\link{SingleCellExperiment-class}} object containing the expression data (raw counts, logcounts and metadata) for all network genes. Required metadata columns:
\describe{
\item{species}{Character, the name of the species.}
\item{clone}{Character, the name of the clone/cell line.}
}}

\item{source}{Character or character vector specifying the source of known transcriptional regulators. It can be one or more of "jaspar_core" (JASPAR 2024 vertebrate core, the default), "jaspar_unvalidated" (JASPAR 2024 unvalidated), and "image" (IMAGE database from Madsen et al. 2018), or alternatively, it can be a  custom list of regulators provided by the user.}

\item{...}{Additional arguments passed to \code{\link{getTopHVGs}}.}
}
\value{
A character vector of known transcriptional regulators that are highly variable in at least one species of the data.
}
\description{
Selects relevant transcriptonal regulators by intersecting highly variable genes with known transcriptional regulators from the specified source(s). These regulators can be used in subsequent steps of the CroCoNet pipeline to assemble modules around them.
}
\details{
This approach combines information from the dataset at hand with prior biological knowledge to pinpoint transcriptional regulators that could be interesting to investigate within the regulatory networks.

As for the information from the dataset, the function identifies highly variable genes in the data using \code{\link{modelGeneVar}} and \code{\link{getTopHVGs}}. Briefly, for each species a trend is fitted between the variance and mean of the log-expression values across all genes, and the fitted value of a gene is regarded as the technial component of variation, while the residual from the trend is regarded as the biological component of variation. All genes with a positive biological component are selected as highly variable genes in each species. If more/less lenient criteria are desired, further arguments can be passed to \code{\link{getTopHVGs}} (e.g. \code{n}, \code{prop}, \code{var.threshold}, or \code{fdr.threshold}). As the final set of highly variable genes, the union of highly variable genes is taken across species.

As for the prior biological knowledge, the function extracts a list of known transcriptional regulators that have at least 1 annotated motif in the motif database(s) specified in \code{source}. This database can be JASPAR 2024 vertebrate core ("jaspar_core"), JASPAR 2024 unvalidated ("jaspar_unvalidated"), the IMAGE database from Madsen et al. 2018 ("image"), or any combination of the above. Alternatively, the user can provide a custom list of regulators selected by their preferred method.

Finally, the highly variable genes and known transcriptional regulators are intersected to provide the final list of regulators that are relevant for the biological process at hand. In the next steps of the pipeline, co-expression modules are assigned around each of these regulators (see \code{\link{assignInitialModules}} and \code{\link{pruneModules}}), which can then be studied in terms of their cross-species conservation.
}
\examples{
regulators <- getRegulators(sce)
regulators2 <- getRegulators(sce, n = 2000)
regulators3 <- getRegulators(sce, c("jaspar_core", "jaspar_unvalidated", "image"))
}
