% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTreeDf}
\alias{getTreeDf}
\title{Convert a tree to a data frame}
\usage{
getTreeDf(tree)
}
\arguments{
\item{tree}{Object of class [phylo] with clones as tips. It is expected to have a component 'species' that specifies which species each tip belongs to.}
}
\value{
A data frame of tree branches.
}
\description{
Converts a tree to a data frame where each row corresponds to a branch in the tree.
}
