% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{image_TRs}
\alias{image_TRs}
\title{IMAGE transcriptional regulators}
\format{
A character vector of 1351 elements.
}
\usage{
image_TRs
}
\description{
Transcriptional regulators that have at least 1 annotated motif in the IMAGE database (Madsen et al. 2018).
}
\keyword{datasets}
