% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initial_modules}
\alias{initial_modules}
\title{Initial modules}
\format{
A data frame with 1750 rows and 8 columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{target}{Target gene of the transcriptional regulator (member of the regulator's initial module).}
\item{weight}{Consensus edge weight/adjacency, the weighted average of clonewise adjacencies.}
\item{n_supporting_clones}{The number of clones where the edge was detected.}
\item{supporting_clones}{The list of clones where the edge was detected.}
\item{rho}{Approximate Spearman's correlation coefficient of the 2 genes' expression profiles that form the edge.}
\item{p.adj}{BH-corrected approximate p-value of rho.}
\item{direction}{Direction of the interaction between the 2 genes that form the edge ("+" or "-").}
}
}
\usage{
initial_modules
}
\description{
Initial modules created by assigning the top 250 targets to each of 7 transcriptional regulators involved in the early neuronal differentiation of primates (\code{regulators}). For each regulator, the top 250 targets were selected based edge weights in the consensus network: all targets of the regulator were ranked based on their edge weight to the regulator (regulator-taregt adjacency) and the 250 targets with the highest regulator-target adjacencies were kept.
}
\keyword{datasets}
