% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jaspar_core_TRs}
\alias{jaspar_core_TRs}
\title{JASPAR 2024 vertebrate core transcriptional regulators}
\format{
A character vector of 784 elements.
}
\usage{
jaspar_core_TRs
}
\description{
Transcriptional regulators that have at least 1 annotated motif in the JASPAR 2024 vertebrate core collection.
}
\keyword{datasets}
