% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jaspar_unvalidated_TRs}
\alias{jaspar_unvalidated_TRs}
\title{JASPAR 2024 unvalidated transcriptional regulators}
\format{
A character vector of 590 elements.
}
\usage{
jaspar_unvalidated_TRs
}
\description{
Transcriptional regulators that have at least 1 annotated motif in the JASPAR 2024 unvalidated collection.
}
\keyword{datasets}
