% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGRNBoost2output.R
\name{loadGRNBoost2output}
\alias{loadGRNBoost2output}
\title{Load the GRNBoost2 output files as a list of igraph objects}
\usage{
loadGRNBoost2output(
  path,
  clone_names = NULL,
  rep = NULL,
  min_occurrence = 2L,
  n_cores = 1L
)
}
\arguments{
\item{path}{Path where the GRNBoost2 output files can be found.}

\item{clone_names}{The names of the clones/cell lines that were input for network reconstruction. Expected to be the base names of the GRNBoost2 TSV files.}

\item{rep}{The number of GRNBoost2 output files per dataset, i.e. the number of independent runs. The GRNBoost2 TSV files are expected to be indexed from 1 to 'rep' for each dataset.}

\item{min_occurrence}{The minimum number of occurrences an edges has to have across runs and directions to be kept (the highest possible number of occurrences is 2×(# of runs)).}

\item{n_cores}{Number of cores.}
}
\value{
A named list that contains the networks of each clone in an igraph format, with edge weights averaged across different runs and directions, and rarely occurring edges removed. The list names are taken from the argument \code{clone_names}.
}
\description{
Loads the TSV files that the network inference algorithm GRNBoost2 outputs and summarizes them as a list of igraph objects.
If GRNBoost2 was run several times on the same data to circumvent the stochastic nature of the algorithm, the edge weights are averaged across the different runs, and a single combined igraph object is returned per clone. The TSV files corresponding to the same clone but different runs are expected to have the same base name: nameOfClone_indexOfRun.tsv.
In addition, edges inferred between the same gene pair but in opposite directions are also averaged. Rarely occurring edges can be removed altogether by specifying 'min_occurrence'.
}
\examples{
network_list_raw <- loadGRNBoost2output(system.file("extdata", package = "CroCoNet"))
}
