% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_list}
\alias{network_list}
\title{List of scaled and filtered networks with edge directionality}
\format{
A named list of 9 [igraph] objects. Each network contains 300 nodes, and has 1 node attribute and 6 edge attributes.

Node attributes:
\describe{
\item{name}{Name of the node (gene).}
}
Edge attributes:
\describe{
\item{weight}{Edge weight, the importance score calculate by GRNBoost2 rescaled between 0 and 1.}
\item{n_supporting_edges}{The number of times an edge appears across 10 independent GRNBoost2 runs and the 2 possible directions (maximum value: 10 runs * 2 direction = 20).}
\item{distance}{Genomic distance of the 2 genes that form the edge. If the genes are on different chromosomes/contigs, the distance is Inf.}
\item{rho}{Approximate Spearman's correlation coefficient of the 2 genes' expression profiles that form the edge.}
\item{p.adj}{BH-corrected approximate p-value of rho.}
\item{direction}{Direction of the interaction between the 2 genes that form the edge ("+" or "-").}
}
}
\usage{
network_list
}
\description{
List of networks per clone with edge weights re-scaled between 0 and 1, gene pairs (edges) with overlapping annotations removed, and the directionality of each edge calculated. The networks were inferred using GRNBoost2 based on a subset of the primate neural differentiation scRNA-seq dataset. All 300 genes in the subsetted data were used as potential regulators. To circumvent the stochastic nature of the algorithm, GRNBoost2 was run 10 times on the same count matrices, then the results were averaged across runs, and rarely occurring edges were removed altogether. In addition, edges inferred between the same gene pair but in opposite directions were also averaged. Edge weights were scaled by the maximum edge weight across all clones. Gene pairs that have overlapping annotations in any of the species' genomes were removed from all networks. The directionality of each edge was determined based on a modified Spearman's correlation between the corresponding 2 genes' expression profiles (positive expression correlation - activating interaction, negative expression correlation - repressing interaction).
}
\keyword{datasets}
