% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_list_scaled}
\alias{network_list_scaled}
\title{List of scaled networks}
\format{
A named list of 9 [igraph] objects. Each network contains 300 nodes, and has 1 node and 2 edge attributes:

Node attributes:
\describe{
\item{name}{Name of the node (gene).}
}
Edge attributes:
\describe{
\item{weight}{Edge weight, the importance score calculate by GRNBoost2 rescaled between 0 and 1.}
\item{n_supporting_edges}{The number of times an edge appears across 10 independent GRNBoost2 runs and the 2 possible directions (maximum value: 10 runs * 2 direction = 20).}
}
}
\usage{
network_list_scaled
}
\description{
List of networks per clone with edge weights re-scaled between 0 and 1. The networks were inferred using GRNBoost2 based on a subset of the primate neural differentiation scRNA-seq dataset. All 300 genes in the subsetted data were used as potential regulators. To circumvent the stochastic nature of the algorithm, GRNBoost2 was run 10 times on the same count matrices, then the results were averaged across runs, and rarely occurring edges were removed altogether. In addition, edges inferred between the same gene pair but in opposite directions were also averaged. Edge weights were scaled by the maximum edge weight across all clones.
}
\keyword{datasets}
