% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpr.R
\name{plotEigengeneHeatmap}
\alias{plotEigengeneHeatmap}
\title{Plot module eigengenes as a heatmap}
\usage{
plotEigengeneHeatmap(
  eigengenes,
  expr_column = "eigengene",
  order_by = "pseudotime",
  heatmap_colors = NULL,
  annotation_colors = NULL,
  font_size = 14
)
}
\arguments{
\item{eigengenes}{Data frame of eigengenes, required columns:
\describe{
\item{cell}{Character, the cell barcode.}
\item{\{\{order_by\}\}}{Numeric or character, the column by which the cells should be ordered on the plot. Typically contains the inferred pseudotime, or the cell type labels.}
\item{module}{Character, transcriptional regulator and in case the eigengene was calculated for the positively or negatively regulated targets only, the direction of regulation (format: nameOfRegulator(+) or nameOfRegulator(-)).}
\item{\{\{expr_column\}\}}{Numeric, summarized module expression profiles (typically the eigengene, the mean expression of the module, or the expression of the regulator).}
}}

\item{expr_column}{Character specifying the column of \code{eigengenes} by which the heatmap should be colored. This column is expected to contain summarized module expression profiles, typically the eigengene (default: "eigengene"), the mean expression of the module, or the expression of the regulator.}

\item{order_by}{Character specifying which column of \code{eigengenes} by which the cells should be ordered on the plot. This column typically contains the inferred pseudotime (default: "pseudotime"), or the cell type labels.}

\item{heatmap_colors}{Character vector, the heatmap colors for the variable specified by \code{expr_column}. The vector can contain any number of colors that will be passed on to and converted into a continuous scale by \code{scale_color_gradientn}.}

\item{annotation_colors}{Character vector, the colors for the variable specified by \code{order_by}. If the variable is discrete, the vector should contain as many colors as there are unique values of the variable, if the variable is continuous, the vector can contain any number of colors that will be passed on to and converted into a continuous scale by \code{scale_color_gradientn}.}

\item{font_size}{Numeric, font size (default: 14).}
}
\value{
A heatmap as a \code{\link{ggplot}} object showing the summarized expression profiles of the modules across all cells.
}
\description{
Plots the eigengenes (or other types of summarized module expression profiles) as a heatmap, with the cells ordered by pseudotime or cell type.
}
\details{
A concept adapted from WGCNA, the eigengene summarizes the expression profile of an entire module, and it is calculated as the first principal component of the module expression data (see also \code{\link{calculateEigengenes}}). Other possible ways of representing the expression profile of a module include the mean expression and the regulator expression.

The function takes a data frame containing any of these summarized module expression profiles as input (normally the output of \code{\link{calculateEigengenes}}) and plots it as a heatmap, with columns corresponding to cells, rows corresponding to modules and colors corresponding to expression levels.

The column of the data frame containing the module expression profiles can be specified by the argument \code{expr_column} (default: "eigengene"). The colors to represent the expression levels can be controlled by the argument \code{heatmap_colors}.

The cells can be order by pseudotime, cell type or any other variable that makes sense in the given context, the name of the column containing the desired cell property can be specified by the argument \code{order_by} (default: "pseudotime"). The colors to represent this cell property can be controlled by the argument \code{annotation_colors}.

If the eigengene (or mean expression) has been calculated for positively and negatively regulated targets separately, then these will appear as separate rows in the heatmap.
}
\examples{
plotEigengeneHeatmap(eigengenes)
}
