% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpr.R
\name{plotExprViolin}
\alias{plotExprViolin}
\title{Plot expression distributions per cell type and species}
\usage{
plotExprViolin(
  genes,
  sce,
  cell_type_column = "cell_type",
  species_colors = NULL,
  font_size = 14
)
}
\arguments{
\item{genes}{Character vector, the names of the genes for which the expression profiles should be plotted.}

\item{sce}{\code{\link{SingleCellExperiment}} object containing the expression data (raw counts, logcounts and metadata) for all network genes. Required metadata columns:
\describe{
\item{species}{Character, the name of the species.}
\item{\{\{cell_type_column\}\}}{Character, cell type annotation.}
}}

\item{cell_type_column}{Character, the name of the cell type annotation column in the metadata of \code{sce}.}

\item{species_colors}{Character vector, colors per species.}

\item{font_size}{Numeric, font size (default: 14).}
}
\value{
A violin plot as a \code{\link{ggplot}} object showing the expression distributions of the input genes per cell type and species.
}
\description{
Plots the expression distributions of one or more genes per cell type and species, and thus allows the expression patterns to be visually compared across species.
}
\details{
The function produces a violin plot of expression levels per cell type and species, faceted by gene. The colors for the species can be controlled by the parameter \code{species_colors}.

The species and cell type information are taken from the metadata slot of the input \code{sce} object, and the expression data are taken from the logcounts assay of the input \code{sce} object.
}
\examples{
plotExprViolin(c("POU5F1","DNMT3B","TERF1","TDGF1","L1TD1","VIM","MAP1B","MARCKS","PTN","CDH2"),
                         sce)
}
\seealso{
Other functions to plot gene expression profiles: 
\code{\link{plotExprAlongPseudotime}()},
\code{\link{plotExprHeatmap}()},
\code{\link{plotSumExprHeatmap}()},
\code{\link{plotSumExprLine}()}
}
\concept{functions to plot gene expression profiles}
