% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModuleSizeDistribution.R
\name{plotModuleSizeDistribution}
\alias{plotModuleSizeDistribution}
\title{Plot the distribution of module sizes}
\usage{
plotModuleSizeDistribution(modules, font_size = 14)
}
\arguments{
\item{modules}{Data frame of the pruned modules, required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Module size, the numer of target genes assigned to a regulator.}
}}

\item{font_size}{Numeric, font size (default: 14).}
}
\value{
A \code{\link{ggplot}} object showing a histogram of module sizes.
}
\description{
Plots the distribution of module sizes (i.e. the number of target genes assigned to each regulator) as a histogram.
}
\examples{
plotModuleSizeDistribution(pruned_modules)
}
