% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPresStats.R
\name{plotPresStatDistributions}
\alias{plotPresStatDistributions}
\title{Plot the distributions of preservation statistics}
\usage{
plotPresStatDistributions(
  pres_stats,
  random_pres_stats,
  stats,
  colors = NULL,
  font_size = 14
)
}
\arguments{
\item{pres_stats}{Data frame of the preservation statistics for the actual (pruned) modules. Required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{clone1, clone2}{Character, the names of the clones compared.}
\item{species1, species2}{Character, the names of the species \code{clone1} and \code{clone2} belongs to, respectively.}
\item{\{\{nameOfStat\}\}}{Numeric, one or more columns containing the preservation statistic(s) of interest per module and clone pair.}
}}

\item{random_pres_stats}{Data frame of the preservation statistics for the random modules. Required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{clone1, clone2}{The names of the clones compared.}
\item{species1, species2}{The names of the species \code{clone1} and \code{clone2} belongs to, respectively.}
\item{\{\{nameOfStat\}\}}{Numeric, one or more columns containing the preservation statistic(s) of interest per module and clone pair.}
}}

\item{stats}{Character or character vector, the name(s) of the column(s) containing the statistic(s) of interest.}

\item{colors}{Character vector of length 2, the colors for the actual and random modules.}

\item{font_size}{Numeric, font size (default: 14).}
}
\value{
A violin plot as a \code{\link{ggplot}} object showing the distributions of chosen preservation statistics both within species and across species and both for the actual and for the random modules.
}
\description{
Plots the distribution(s) of the chosen preservation statistic(s) for the actual and random modules and for within-species and cross-species clone pairs. If a statistic is a good measure of module preservation, the preservation scores are expected to be 1) higher for the actual modules than for the random modules and 2) higher within species than across species.
}
\details{
As part of the CroCoNet approach, pairwise module preservation scores are calculated between clones, both within and across species (see \code{\link{calculatePresStats}}) to gain information about the cross-species differences but also about the within-species diversity of the modules. These correlation-based preservation statistics quantify how well the module topology is preserved between the networks of two clones. The statistics are calculated not just for the actual, biologically meaningful modules, but also for random modules with matching sizes.

The actual modules are expected to be more preserved than the random modules, and all modules, but especially the actual ones, are expected to be more preserved within species than across species. The function plots the distributions of the within-species and cross-species scores both for the actual and for the random modules and thus allows these expectations to be visually checked. If several different statistics are input, these will be shown as the rows of the faceted plot. When it comes to choosing the best statistic, it is recommended to take the one that follows these expected trends the most.
}
\examples{
plotPresStatDistributions(pres_stats, random_pres_stats, "cor_kIM")
}
\seealso{
Other functions to plot preservation statistics: 
\code{\link{plotPresStats}()}
}
\concept{functions to plot preservation statistics}
