% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPresStats.R
\name{plotPresStats}
\alias{plotPresStats}
\title{Plot cross-species VS within-species preservation statistics per species pair}
\usage{
plotPresStats(
  pres_stats,
  stats,
  colors = NULL,
  font_size = 14,
  point_size = 0.3,
  point_alpha = 0.7
)
}
\arguments{
\item{pres_stats}{Data frame of the preservation statistics for the actual (pruned) modules. Required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Integer, the numer of target genes assigned to a regulator.}
\item{clone1, clone2}{Character, the names of the clones compared.}
\item{species1, species2}{Character, the names of the species 'clone1' and 'clone2' belongs to, respectively.}
\item{\{\{nameOfStat\}\}}{Numeric, one or more columns containing the preservation statistic of interest per module and clone pair.}
}}

\item{stats}{Character or character vector, the name(s) of the column(s) containing the statistics of interest.}

\item{colors}{Character vector, the colors to visualize the module sizes. The vector can contain any number of colors that will be passed on to and converted into a continuous scale by \code{scale_color_gradientn}.}

\item{font_size}{Numeric, font size (default: 14).}

\item{point_size}{Numeric, the size of the points (default: 0.3).}

\item{point_alpha}{Numeric, the opacity of the points (default: 0.7).}
}
\value{
A \code{\link{ggplot}} object.
}
\description{
Plots the cross-species VS within-species preservation scores for all possible species pairs. If a statistic is a good measure of module preservation, the within-species scores are expected to be higher than the cross-species scores for all pairs, and the difference is expected to increase with increasing phylogenetic distance between the two species compared.
}
\details{
As part of the CroCoNet approach, pairwise module preservation scores are calculated between clones, both within and across species (see \code{\link{calculatePresStats}}) to gain information about the cross-species differences but also about the within-species diversity of the modules. These correlation-based preservation statistics quantify how well the module connectivity patterns are preserved between the networks of two clones.

For each possible speciesA-speciesB pair in the data, the function first subsets the within-species scores for speciesA, the within-species scores for speciesB and the cross-species scores between speciesA and speciesB, then calculates the mean of the within-species scores as well as the mean of the cross-species scores per module. For example, if there are 2 species, human and gorilla, with 3 human clones (\emph{h1}, \emph{h2} and \emph{h3}), and 2 gorilla clones (\emph{g1} and \emph{g2}), for a single module there will be 3 preservation scores within the human clones (\eqn{p_{h1-h2}}, \eqn{p_{h1-h3}} and \eqn{p_{h2-h3}}), 1 preservation score within the gorilla clones (\eqn{p_{g1-g2}}) and 6 preservation scores across the 2 species (\eqn{p_{h1-g1}}, \eqn{p_{h1-g2}}, \eqn{p_{h2-g1}}, \eqn{p_{h2-g2}}, \eqn{p_{h3-g1}} and \eqn{p_{h3-g2}}). For this module and the human-gorilla species pair, the summarized within-species score will then be the mean of the 4 within-species scores (\eqn{mean(p_{h1-h2}, p_{h1-h3}, p_{h2-h3}, p_{g1-g2})}) and the cross-species score will be the mean of the 6 cross-species scores (\eqn{mean(p_{h1-g1}, p_{h1-g2}, p_{h2-g1}, p_{h2-g2}, p_{h3-g1}, p_{h3-g2})}).

After calculating the summarized within-species and cross-species scores for all modules and species pairs, the data points are shown faceted by species pair and colored by module size. The within-species scores are expected to be higher than the cross-species scores, thus the data points are expected to fall under the diagonal in all facets. However, the more phylogenetically distant two species are, the bigger deviation we expect from the diagonal. If several different statistics are input, these will be shown as the rows of the faceted plot. When it comes to choosing the best statistic, it is recommended to take the one that follow the most these expected trends.
}
\examples{
plotPresStats(pres_stats, "cor_kIM")
}
