% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpr.R
\name{plotSumExprHeatmap}
\alias{plotSumExprHeatmap}
\title{Plot summarized expression profiles as a heatmap}
\usage{
plotSumExprHeatmap(
  genes,
  sce,
  summary_metric = "pct",
  cell_type_column = "cell_type",
  heatmap_colors = NULL,
  species_colors = NULL,
  cell_type_colors = NULL,
  clip = 3,
  font_size = 14,
  label_size = font_size/3
)
}
\arguments{
\item{genes}{Character vector, the names of the genes for which the expression profiles should be plotted.}

\item{sce}{\code{\link{SingleCellExperiment}} object containing the expression data (logcounts and metadata) for all network genes. Required metadata column:
\describe{
\item{species}{Character, the name of the species.}
\item{\{\{cell_type_column\}\}}{Character, cell type annotation.}
}}

\item{summary_metric}{Character, the metric for summarizing the expression of a gene per cell type and species. One of "pct" (the percent of cells expressing the gene, the default) and "mean" (mean expression in logcounts).}

\item{cell_type_column}{Character, the name of the cell type annotation column in the metadata of \code{sce} (default: "cell_type").}

\item{heatmap_colors}{Character vector, the heatmap colors for the summarized expression levels. The vector can contain any number of colors that will be passed on to and converted into a continuous scale by \code{\link{scale_color_gradientn}}.}

\item{species_colors}{Character vector, colors per species.}

\item{cell_type_colors}{Character vector, colors per cell type.}

\item{clip}{Numeric specifying the degree of clipping. For each gene, the expression level values that are more standard deviations away from the mean than \code{clip} are treated as NA. The default is 3 meaning that expression levels are clipped to the range of mean \eqn{\pm} 3\eqn{\sigma} per gene.}

\item{font_size}{Numeric, font size (default: 14).}

\item{label_size}{Numeric, font size of the cell type labels (default: \code{font_size} / 5).}
}
\value{
A heatmap as a \code{\link{ggplot}} object showing the summarized expression levels of the input genes per cell type and species.
}
\description{
Plots the summarized expression of one or more genes per cell type and species as a heatmap.
}
\details{
The function creates a heatmap, with columns corresponding to cell types, rows corresponding to genes and species, and colors corresponding to either percent of cells expressing the given gene (if \code{summary_metric} = "pct", the default) or mean expression in logcounts (if \code{summary_metric} = "mean").

The cell types and species are visualized as annotation graphics above the heatmap using \code{cell_type_colors} and on the right hand side of the heatmap using \code{species_colors}, respectively.

Expression levels are clipped to the range of mean \eqn{\pm} 3\eqn{\sigma} per gene, this can be changed via the parameter \code{clip}. Clipping aids visualization by preventing the outlier data points from squishing the rest of the data into a small color range. If clipping is not desired, please set \code{clip} to Inf.
}
\examples{
plotSumExprHeatmap(c("POU5F1","DNMT3B","TERF1","TDGF1","L1TD1","VIM","MAP1B","MARCKS","PTN","CDH2"),
                             sce)
}
\seealso{
Other functions to plot gene expression profiles: 
\code{\link{plotExprAlongPseudotime}()},
\code{\link{plotExprHeatmap}()},
\code{\link{plotExprViolin}()},
\code{\link{plotSumExprLine}()}
}
\concept{functions to plot gene expression profiles}
