% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExpr.R
\name{plotSumExprLine}
\alias{plotSumExprLine}
\title{Plot summarized expression profiles as a line plot}
\usage{
plotSumExprLine(
  genes,
  sce,
  summary_metric = "pct",
  cell_type_column = "cell_type",
  species_colors = NULL,
  cell_type_colors = NULL,
  font_size = 14
)
}
\arguments{
\item{genes}{Character vector, the names of the genes for which the expression profiles should be plotted.}

\item{sce}{\code{\link{SingleCellExperiment}} object containing the expression data (raw counts, logcounts and metadata) for all network genes. Required metadata columns:
\describe{
\item{species}{Character, the name of the species.}
\item{\{\{cell_type_column\}\}}{Character, cell type annotation.}
}}

\item{summary_metric}{Character, the metric for summarizing the expression of a gene per cell type and species. One of "pct" (the percent of cells expressing the gene, the default) and "mean" (mean expression in logcounts).}

\item{cell_type_column}{Character, the name of the cell type annotation column in the metadata of \code{sce}.}

\item{species_colors}{Character vector, colors per species.}

\item{cell_type_colors}{Character vector, colors per cell type.}

\item{font_size}{Numeric, font size (default: 14).}
}
\value{
A line plot as a \code{\link{ggplot}} object showing the summarized expression levels of the genes per cell type and species.
}
\description{
Plots the summarized expression of one or more genes per cell type and species as a line plot, and thus allows the expression patterns to be visually compared across species.
}
\details{
The expression of each input gene is summarized either as the percent of cells expressing the given gene (if \code{summary_metric} = "pct", the default) or the mean expression and its 95% confidence interval in logcounts (if \code{summary_metric} = "mean") per cell type and species.#' The species and cell type information are taken from the metadata slot of the input \code{sce} object, and the expression data are taken from the logcounts assay of the input \code{sce} object.

The summarized expression levels are then visualized as a scatter plot across the cell types, colored by species and faceted by gene. To aid visual comparison across species, the data points of the same species are connected by lines across the cell types (this is not meant to represent continuity).

The cell types are depicted below the plot using \code{cell_type_colors}.
}
\examples{
plotSumExprLine(c("POU5F1","DNMT3B","TERF1","TDGF1","L1TD1","VIM","MAP1B","MARCKS","PTN","CDH2"),
                          sce)
}
\seealso{
Other functions to plot gene expression profiles: 
\code{\link{plotExprAlongPseudotime}()},
\code{\link{plotExprHeatmap}()},
\code{\link{plotExprViolin}()},
\code{\link{plotSumExprHeatmap}()}
}
\concept{functions to plot gene expression profiles}
