% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pruned_modules}
\alias{pruned_modules}
\title{Pruned modules}
\format{
A data frame with 225 rows and 9 columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{target}{Character, target gene of the transcriptional regulator (member of the regulator's pruned module).}
\item{weight}{Numeric, consensus edge weight/adjacency, the weighted average of clonewise adjacencies.}
\item{n_supporting_clones}{The number of clones where the edge was detected.}
\item{supporting_clones}{The list of clones where the edge was detected.}
\item{rho}{Approximate Spearman's correlation coefficient of the 2 genes' expression profiles that form the edge.}
\item{p.adj}{BH-corrected approximate p-value of rho.}
\item{direction}{Character specifying the direction of regulation between the regulator and the target, either "+" or "-".}
\item{module_size}{Module size, the numer of target genes assigned to a regulator.}
}
}
\usage{
pruned_modules
}
\description{
Pruned modules created by the dynamic pruning of the initial modules via the "UIK_adj_kIM" method. The initial module members were filtered in successive steps based on their adjacency to the regulator and their intramodular connectivitiy alternately. In each step, the cumulative sum curve based on one of these two characteristics was calculated per module, then the targets below the knee point of the curve were kept. This process was continued until the module sizes became as small as possible without the median falling below a pre-defined minimum of 20 genes.
}
\keyword{datasets}
