% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneModules.R
\name{pruneModules_UIK_adj}
\alias{pruneModules_UIK_adj}
\title{Prune modules based on the regulator-target adjacencies using dynamic filtering}
\usage{
pruneModules_UIK_adj(
  initial_modules,
  min_module_size = 20L,
  min_median_module_size = 20L,
  max_frac_modules_lost = 0.02,
  exponent = 1L
)
}
\arguments{
\item{initial_modules}{Data frame of the initial modules with columns 'regulator', 'target' and 'weight' containing the transcriptional regulators, their target genes and the edge weights between each regulator-target pair.}

\item{min_module_size}{Integer, the lower treshold of module size. Modules with a smaller size than this threshold are removed after each pruning step.}

\item{min_median_module_size}{Integer, the lower threshold of median module size.}

\item{max_frac_modules_lost}{Numeric, the upper threshold of the fraction of removed modules.}

\item{exponent}{Integer, the exponent the regulator-target adjacency and intramodular connectivity is raised to the power of during the cumulative sum curve calculation.
The initial module members are filtered in successive steps based on their adjacency to the regulator. In each step, the cumulative sum curves of the regulator-target adjacencies are calculated per module, the knee point of the curve is identified using the Unit Invariant Knee (UIK) method, then only the targets that rank higher than the knee point are kept. If 'exponent' is set higher than 1, the adjacencies are raised to the equivalent power when calculating the cumulative sum curves and its knee point. The modules containing less target genes than 'min_module_size' are removed after each pruning step. The steps continue until the median module size becomes as small as possible without falling below 'min_median_module_size' OR until the fraction of removed modules becomes higher than 'max_frac_modules_lost'.}
}
\value{
Data frame of the pruned modules.
}
\description{
Prunes the initial modules by applying a dynamic stepwise pruning based on the regulator-target adjacencies.
}
\examples{
pruned_modules_UIK_adj <- pruneModules_UIK_adj(initial_modules)
}
