% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_modules}
\alias{random_modules}
\title{Random modules}
\format{
A data frame with 225 rows and 3 columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{target}{Member of the regulator's random module.}
\item{module_size}{Module size, the numer of target genes assigned to a regulator.}
}
}
\usage{
random_modules
}
\description{
Random modules that match the actual (pruned) modules in size. These random modules have the same regulators and contain the same number of target genes as the actual modules, but the target genes were randomly drawn from all genes in the network.
}
\keyword{datasets}
