% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_tree_stats_jk}
\alias{random_tree_stats_jk}
\title{Tree-based statistics of the original and jackknifed random modules}
\format{
A data frame with 232 rows and 16 columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Module size, the numer of target genes assigned to a regulator.}
\item{type}{Character, module type (orig = original or jk = jackknifed).}
\item{id}{Character, the unique ID of the module version (format: nameOfRegulator_jk_nameOfGeneRemoved in case of module type 'jk' and nameOfRegulator_orig in case of module type 'orig').}
\item{gene_removed}{The name of the gene removed by jackknifing (NA in case of module type 'orig').}
\item{within_species_diversity}{Numeric, the sum of the human, gorilla and cynomolgus diversity.}
\item{total_tree_length}{Numeric, the sum of all branch lengths in the tree.}
\item{human_diversity}{Numeric, the sum of the branch lengths in the subtree that contains only the human tips.}
\item{gorilla_diversity}{Numeric, the sum of the branch lengths in the subtree that contains only the gorilla tips.}
\item{cynomolgus_diversity}{Numeric, the sum of the branch lengths in the subtree that contains only the cynomolgus tips.}
\item{human_monophyl}{Logical indicating whether the tree is monophyletic for the human clones.}
\item{gorilla_monophyl}{Logical indicating whether the tree is monophyletic for the gorilla clones.}
\item{cynomolgus_monophyl}{Logical indicating whether the tree is monophyletic for the cynomolgus clones.}
\item{human_to_other_branch_length}{Numeric, the length of the internal branch that connects the subtree of the human clones to the subtree of all other clones. NA if the tree is not monophyletic for the human clones.}
\item{gorilla_to_other_branch_length}{Numeric, the length of the internal branch that connects the subtree of the gorilla clones to the subtree of all other clones. NA if the tree is not monophyletic for the gorilla clones.}
\item{cynomolgus_to_other_branch_length}{Numeric, the length of the internal branch that connects the subtree of the cynomolgus clones to the subtree of all other clones. NA if the tree is not monophyletic for the cynomolgus clones.}
}
}
\usage{
random_tree_stats_jk
}
\description{
Tree-based statistics calculated for the original and all jackknifed versions of the random modules. For each module version, a tree was reconstructed based on the module preservation scores within and across species. The tips of resulting tree represent the clones and the branch lengths represent the dissimilarity of connectivity patterns between the clones. Branches between clones of different species carry information about cross-species differences, while the branches between clones of the same species carry information about the within-species diversity.
}
\keyword{datasets}
