% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_trees_jk}
\alias{random_trees_jk}
\title{Trees of the original and jackknifed random modules}
\format{
A named list with 232 elements containing the neighbor-joining trees as [phylo] objects.
}
\usage{
random_trees_jk
}
\description{
Neighbor-joining trees representing the similarities of connectivity patterns across the 9 primate replicates for the original and all jackknifed versions of the random modules. The jackknifed versions of the modules were created by removing each target gene assigned to a module (the regulators were never excluded). For each of these module versions, the trees were inferred based on the preservation statistic cor.kIM (correlation of intramodular connectivities): first, the preservation scores were calculated between all possible replicate pairs, then they were converted into a distance matrix of replicates, and finally trees were reconstructed based on this distance matrix using the neighbor-joining algorithm. The result is a single tree for each original or jackknife module where the tips represent the replicates and the branch lengths represent the dissimilarity of connectivity patterns between these replicates.
}
\keyword{datasets}
