% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTrees.R
\name{reconstructTrees}
\alias{reconstructTrees}
\title{Reconstruct trees across all modules}
\usage{
reconstructTrees(dist_list, n_cores = 1L)
}
\arguments{
\item{dist_list}{A named list containing the distance measures as data frames for all modules or jackknifed module versions. Required columns for the data frames:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Integer, the numer of target genes assigned to a regulator.}
\item{type}{Character, module type (orig = original or jk = jackknifed, only needed if the distances were calculated with jackknifing).}
\item{id}{Character, the unique ID of the module version (format: nameOfRegulator_jk_nameOfGeneRemoved in case of module type 'jk' and nameOfRegulator_orig in case of module type 'orig', only needed if the distances were calculated with jackknifing).}
\item{gene_removed}{Character, the name of the gene removed by jackknifing (NA in case of module type 'orig', only needed if the distances were calculated with jackknifing).}
\item{clone1, clone2}{Character the names of the clones compared.}
\item{species1, species2}{Character, the names of the species \code{clone1} and \code{clone2} belong to, respectively.}
\item{dist}{Numeric, the distance measure to be used for the tree reconstruction.}
}}

\item{n_cores}{Number of cores.}
}
\value{
A named list containing the neighbor-joining trees as \code{\link{phylo}} objects for all modules or jackknifed module versions in the input \code{dist_list}.
}
\description{
Reconstructs neighbor-joining trees based on pairwise distance measures between clones.
}
\examples{
trees_jk <- reconstructTrees(dist_jk)
}
