% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulators}
\alias{regulators}
\title{Transcriptional regulators}
\format{
A character vector of 7 elements.
}
\usage{
regulators
}
\description{
7 important transcriptional regulators during early neural differentiation that are used as cores to assemble modules around them.
}
\keyword{datasets}
