% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOverlappingGenePairs.R
\name{removeOverlappingGenePairs}
\alias{removeOverlappingGenePairs}
\title{Remove gene pairs with overlapping annotations}
\usage{
removeOverlappingGenePairs(
  network_list,
  gtf_list,
  clone2species,
  gene_col = c("gene_name", "gene_id"),
  n_cores = 1L
)
}
\arguments{
\item{network_list}{A named list that contains the networks of each clone in an igraph format.}

\item{gtf_list}{A named list that contains the genome annotation of each species in a GenomicRanges format.}

\item{clone2species}{A data frame with columns 'clone' and 'species' that specifies which species each clone belongs to. The names of clones and species should match the names of 'network_list' and gtf_list'.}

\item{gene_col}{One of 'gene_name' or 'gene_id', specifies what type of identifier the network nodes have.}

\item{n_cores}{Number of cores.}
}
\value{
A named list of networks after the removal of gene pairs with overlapping annotations.
}
\description{
Removes gene pairs that have overlapping annotations in any of the species' genomes from all networks.
}
\examples{
network_list_scaled_filt <- removeOverlappingGenePairs(network_list_scaled,
                                                       gtf_list,
                                                       clone2species,
                                                       "gene_name")
}
