% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{replicate2species}
\alias{replicate2species}
\title{Replicate-species conversion}
\format{
A data frame with 9 rows and 2 columns:
\describe{
\item{replicate}{Name of the replicate/cell line.}
\item{species}{Name of the species.}
}
}
\usage{
replicate2species
}
\description{
A data frame that specifies which replicate belongs to which species.
}
\keyword{datasets}
