% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleEdgeWeights.R
\name{rescaleEdgeWeights}
\alias{rescaleEdgeWeights}
\title{Rescale edge weights between 0 and 1}
\usage{
rescaleEdgeWeights(network_list, n_cores = 1L)
}
\arguments{
\item{network_list}{A named list that contains the networks of each clone in an igraph format.}

\item{n_cores}{Number of cores.}
}
\value{
A named list that contains the networks of each clone in an igraph format, with the edge weights rescaled between 0 and 1.
}
\description{
Rescale edge weights between 0 and 1
}
\examples{
network_list_scaled <- rescaleEdgeWeights(network_list_raw)
}
