% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free_transformation.R
\name{scaleFreeFitRSquared}
\alias{scaleFreeFitRSquared}
\title{Calculate R2 value of the scale-free model fit for the provided connectivities}
\usage{
scaleFreeFitRSquared(k)
}
\arguments{
\item{k}{Numeric vector, the connectivities of a network.}
}
\value{
Numeric, the \eqn{R^2} value of the scale-free model fit.
}
\description{
Calculate R2 value of the scale-free model fit for the provided connectivities
}
