% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_list}
\alias{sce_list}
\title{SCE object list of the primate neural differentiation dataset}
\format{
A named list of 9 SCE objects. Each SCE object has 300 rows, 100 columns, 9 metadata columns and 3 assays.

Metadata columns:
\describe{
\item{species}{Name of the species.}
\item{clone}{Name of the clone/cell line.}
\item{day}{The day when the cell was collected.}
\item{n_UMIs}{Number of UMIs detected.}
\item{n_genes}{Number of genes detected.}
\item{perc_mito}{Percent of mitochondrial reads.}
\item{sizeFactor}{Size factor for scaling normalization, calculated per clone using [scran::computeSumFactors] and [scran::quickCluster].}
\item{pseudotime}{Pseudotime inferred by [SCORPIUS::infer_trajectory].}
\item{cell_type}{Cell type labels predicted by [SingleR::classifySingleR] using the embryoid body dataset from Rhodes et al. 2022 as reference.}
}
Assays:
\describe{
\item{counts}{Raw counts.}
\item{logcounts}{Log-normalized counts created by calculating size factors per clone using [scran::computeSumFactors] and [scran::quickCluster], log-normalizing by [scuttle::logNormCounts].}
\item{rand_quantile_res}{Randomized quantile residuals calculated by [transformGamPoi::transformGamPoi].}
}
}
\usage{
sce_list
}
\description{
A subset of the primate neural differentiation scRNA-seq dataset as a list of SCE objects. The data was collected during the early neural differentiation of human, gorilla and cynomolgus macaque iPS cells with 3 human, 2 gorilla and 4 cynomolgus cell lines (clones). Using a directed differentiation protocol, cells were differentiated into neural progenitor cells (NPCs) over the course of 9 days, and scRNA-seq data was obtained at six time points (days 0, 1, 3, 5, 7 and 9) during this process. Each element of the list is the SCE object of a clone with 300 genes and 100 cells, containing the raw counts, log-normalized counts and randomized quantile residuals as well as the metadata.
}
\keyword{datasets}
