% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free_transformation.R
\name{selectPower}
\alias{selectPower}
\title{Select power for scale-free transfrmation}
\usage{
selectPower(scale_free_fit, R2_cutoff = 0.85, rescue = TRUE)
}
\arguments{
\item{scale_free_fit}{A data frame containing the \eqn{R^2} values of the scale-free model fit for each network and soft-thresholding power.}

\item{R2_cutoff}{Numeric, the \eqn{R^2} cutoff each network needs to pass when selecting the power (default: 0.85).}

\item{rescue}{Logical indicating whether to still select a power if none of the powers achieve \eqn{R^2} values greater than the specified \code{R2_cut} for all networks. If TRUE (the default), the power with the highest minimum \eqn{R^2} across the networks is output in case the cutoff is not met, if FALSE, the function gives an error in case the cutoff is not met.}
}
\value{
Numeric, the best power to use for the scale-free transformation.
}
\description{
Select power for scale-free transfrmation
}
