% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeJackknifeStats.R
\name{summarizeStat}
\alias{summarizeStat}
\title{Summarize values of a statistic}
\usage{
summarizeStat(values, summary_method, conf_level = 0.95)
}
\arguments{
\item{values}{Numeric, integer or logical vector.}

\item{summary_method}{Character, the measure of central tendency ("mean" or "median") to be used.}

\item{conf_level}{Numeric, confidence level of the interval (default: 0.95).}
}
\value{
A data frame with 1 row and 4 columns:
\describe{
\item{estimate}{Numeric, the central tendency (mean or median) of 'values'.}
\item{var}{Numeric, the variance of 'values'.}
\item{lwr}{Numeric, the lower bound of the confidence interval.}
\item{upr}{Numeric, the upper bound of the confidence interval.}
}
}
\description{
Calculates the estimate (mean or median), the confidence interval of the estimate with the specified confidence level and the variance of the data provided.
}
