% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_free_transformation.R
\name{testPowers}
\alias{testPowers}
\title{Test soft-thresholding power values}
\usage{
testPowers(network_list, powers = 1:20, n_cores = 1L)
}
\arguments{
\item{network_list}{A named list that contains the networks of each clone and the consensus network in an \code{\link{igraph}} format.}

\item{powers}{A numeric or integer vector specifying the soft-thresholding powers to test.}

\item{n_cores}{Integer, the number of cores (default: 1).}
}
\value{
A data frame containing the \eqn{R^2} values of the scale-free model fit for each network and soft-thresholding power.
}
\description{
Applies a range of soft-thresholding powers to the networks provided and determines the goodness of fit for each soft power and network.
}
\examples{
scale_free_fit <- testPowers(network_list, c(1, seq(2, 10, by = 2)))
}
