% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDirectionality.R
\name{addDirectionality}
\alias{addDirectionality}
\title{Add directionality of interactions}
\usage{
addDirectionality(network, sce, assay = "logcounts", n_cores = 1L)
}
\arguments{
\item{network}{An \code{\link{igraph}} object containing the consensus network or the network of a clone.}

\item{sce}{A \code{\link{SingleCellExperiment}} object containing the expression data either for all clones (in case \code{network} is the consensus network) or for the clone of interest (in case \code{network} is a clonewise network). If \code{sce} contains the expression data of all clones, it is also expected to have a metadata column "clone" specifying which clone each cell belongs to; this will be used to define the blocking levels.}

\item{assay}{Character, the name of the assay in \code{sce} that should be used for the calculation of gene-gene correlations (default: "logcounts").}

\item{n_cores}{Integer, the number of cores (default: 1).}
}
\value{
An \code{\link{igraph}} object, the input \code{network} extended by the information about the direction of interactions. In addition to the original edge attributes, it contains 3 new attributes:
\describe{
\item{rho}{Numeric, the approximate Spearman's correlation coefficient between the expression profiles of the 2 genes that form the edge.}
\item{p.adj}{Numeric, BH-corrected approximate p-value of rho.}
\item{direction}{Character, the direction of the interaction between the 2 genes that form the edge ("+" = positively correlated/coexpressed or "-" = negatively correlated/anti-coexpressed).}
}
}
\description{
Determines the directionality of each network edge (positively correlated/coexpressed or negatively correlated/anti-coexpressed) based on a modified Spearman's correlation between the expression profiles of the 2 genes that form the edge.
}
\details{
If the networks were inferred using a method that does not distinguish coexpressed and anti-coexpressed gene pairs, it might be useful to add this information for lines of analysis where it makes sense to separate the activated and repressed target genes of a regulator (e.g. for the calculation of eigengenes, see \code{\link{calculateEigengenes}}). If the network inference method output edges with both positive and negative edge weights in the first place (e.g. correlation-based methods), the edge attribute "direction" is already created during the step \code{\link{normalizeEdgeWeights}} and does not have to be calculated again.

Here the directionality of a geneA-geneB edge refers to the characteristic whether geneA and geneB are coexpressed or anti-coexpressed and NOT whether geneA regulates geneB or geneB regulates geneA. The network remains undirected in a graph theoretical sense.

The calculation of directionality relies on the approximate version of the Spearman's rho, significance testing and blocking implemented by \code{\link{correlatePairs}}. The results are summarized as 3 new edge attributes in the \code{\link{igraph}} object: rho (approximate Spearman's correlation coefficient), p.adj (BH-corrected approximate p-value) and direction ("+" or "-").
}
\examples{
consensus_network <- network_list \%>\%
 createConsensus(clone2species, tree) \%>\%
 addDirectionality(sce)
}
