% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructTrees.R
\name{convertPresToDist}
\alias{convertPresToDist}
\title{Convert preservation statistics to distances across all modules}
\usage{
convertPresToDist(pres_stats, stat, n_cores = 1L)
}
\arguments{
\item{pres_stats}{Data frame of the preservation statistics, required columns:
\describe{
\item{regulator}{Character, transcriptional regulator.}
\item{module_size}{Integer, the numer of target genes assigned to a regulator.}
\item{type}{Character, module type (orig = original or jk = jackknifed, only needed if the preservation statistics were calculated with jackknifing).}
\item{id}{Character, the unique ID of the module version (format: nameOfRegulator_jk_nameOfGeneRemoved in case of module type 'jk' and nameOfRegulator_orig in case of module type 'orig', only needed if the preservation statistics were calculated with jackknifing).}
\item{gene_removed}{Character, the name of the gene removed by jackknifing (NA in case of module type 'orig', only needed if the preservation statistics were calculated with jackknifing).}
\item{clone1, clone2}{Character the names of the clones compared.}
\item{species1, species2}{Character, the names of the species \code{clone1} and \code{clone2} belong to, respectively.}
\item{\{\{stat\}\}}{Numeric, the preservation statistic specified by the argument \code{stat}.}
}}

\item{stat}{Character, the name of the column containing the preservation statistic that is to be converted into a distance measure.}

\item{n_cores}{Number of cores.}
}
\value{
A named list containing the distance measures as data frames for all modules or jackknifed module versions in the input \code{pres_stats}. The data frames contain 1 new column in addition to the relevant columns of \code{pres_stats}:
\describe{
\item{dist}{Numeric, the distance measure ranging from 0 to 1, calculated based on \code{stat}.}
}
}
\description{
Converts pairwise preservation statistics between clones to distance measures ranging from 0 to 1.
}
\examples{
dist_jk <- convertPresToDist(pres_stats_jk, "cor_kIM")
}
