% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createConsensus.R
\name{createConsensus}
\alias{createConsensus}
\title{Create consensus network}
\usage{
createConsensus(network_list, clone2species, tree = NULL)
}
\arguments{
\item{network_list}{A named list that contains the networks of each clone in an igraph format.}

\item{clone2species}{A data frame with columns 'clone' and 'species' that specifies which species each clone belongs to. The names of clones and species should match the names of 'network_list' and the tip labels of 'tree', respectively.}

\item{tree}{Object of class 'phylo' that gives the phylogenetic tree of the species.}
}
\value{
Consensus network in an igraph format.
}
\description{
Integrates networks across different clones and different species into a single consensus network. The input networks should all contain the same nodes (genes). For each edge, the consensus adjacencies are calculated as the weighted average of clonewise adjacencies using weights that correct for 1) the phylogenetic distances between species (if the phylogenetic tree is provided) and 2) the different numbers of clones per species.
}
\examples{
consensus_network <- createConsensus(network_list_scaled_filt, clone2species, tree)
}
