% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CroCoNet-package.R
\docType{package}
\name{CroCoNet-package}
\alias{CroCoNet}
\alias{CroCoNet-package}
\title{CroCoNet: Cross-species Comparison of Networks}
\description{
A tool to quantitatively compare gene regulatory networks across species and identify conserved and diverged modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hellmann-lab.github.io/CroCoNet}
  \item \url{https://hellmann-lab.github.io/CroCoNet/}
  \item Report bugs at \url{https://github.com/Hellmann-Lab/CroCoNet/issues}
}

}
\author{
\strong{Maintainer}: Anita Térmeg \email{anita.termeg@gmail.com} (\href{https://orcid.org/0009-0005-8872-9086}{ORCID})

}
\keyword{internal}
