% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network_list_raw}
\alias{network_list_raw}
\title{List of raw networks}
\format{
A named list of 9 [igraph] objects. Each network contains 300 nodes, and has 1 node and 2 edge attributes:

Node attributes:
\describe{
\item{name}{Name of the node (gene).}
}
Edge attributes:
\describe{
\item{weight}{Edge weight, the importance score calculate by GRNBoost2.}
}
}
\usage{
network_list_raw
}
\description{
List of networks per replicate with raw edge weights. The networks were inferred using GRNBoost2 based on a subset of the primate neural differentiation scRNA-seq dataset. All 300 genes in the subsetted data were used as potential regulators. To circumvent the stochastic nature of the algorithm, GRNBoost2 was run 10 times on the same count matrices, then the results were averaged across runs, and rarely occurring edges were removed altogether. In addition, edges inferred between the same gene pair but in opposite directions were also averaged.
}
\keyword{datasets}
