% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneModules.R
\name{pruneModules_topN}
\alias{pruneModules_topN}
\title{Prune modules based on the regulator-target adjacencies by keeping a fixed number of top targets}
\usage{
pruneModules_topN(initial_modules, N = 50L)
}
\arguments{
\item{initial_modules}{Data frame of the initial modules with columns 'regulator', 'target' and 'weight' containing the transcriptional regulators, their target genes and the edge weights between each regulator-target pair.}

\item{N}{Either an integer specifying a single desired pruned module size for all modules or a named integer vector specifying the desired pruned module size for each regulator.}
}
\value{
Data frame of the pruned modules.
}
\description{
Prunes the initial modules by keeping a fixed number of targets per transcriptional regulator with the highest regulator-target adjacencies
}
\examples{
pruned_modules_top50 <- pruneModules_topN(initial_modules, 50)
}
