% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOverlappingGenePairs.R
\name{removeOverlappingGenePairs}
\alias{removeOverlappingGenePairs}
\title{Remove gene pairs with overlapping annotations}
\usage{
removeOverlappingGenePairs(
  network_list,
  gtf_list,
  clone2species,
  gene_col = c("gene_name", "gene_id"),
  n_cores = 1L
)
}
\arguments{
\item{network_list}{A named list of \code{\link{igraph}} objects containing the networks of all clones.}

\item{gtf_list}{A named list of GRanges objects containing the genome annotations of all species.}

\item{clone2species}{A data frame specifying which species each clone belongs to, required columns:
\describe{
\item{clone}{Character, name of the clone.}
\item{species}{Character, name of the species.}
}}

\item{gene_col}{Character specifying the type of identifier the network nodes have, one of "gene_name", "gene_id". The function looks for the names of the network nodes in the corresponding column of the GTF files.}

\item{n_cores}{Integer, the number of cores (default: 1).}
}
\value{
A named list of \code{\link{igraph}} objects containing the networks of all clones, after the removal of gene pairs with overlapping annotations. A new edge attribute is added to all \code{\link{igraph}} objects:
\describe{
\item{genomic_dist}{Numeric, the genomic distance of the 2 genes that form the edge (\code{Inf} if the 2 genes are annotated on different chromosomes/contigs).}
}
}
\description{
Removes the gene pairs (edges) that have overlapping annotations in the genomes of at least 1 species from all networks.
}
\details{
Mapping and counting is problematic for overlapping genomic features: it is difficult to tell apart which read belongs to which gene. Often parts of the reads from one gene are assigned to the other gene, leading to correlated expression profiles simply due to genomic position. This only has a marginal effect on the results of a DE analysis, but can cause false positive edges with very high edge weights in case of a network analysis.

This functions circumvents such potential artefacts by removing all edges between overlapping genes. As the first step, the positions of the network genes are determined in the genome of each species. Depending on the value of \code{gene_col}, the names of the network nodes are matched to the entries in either the "gene_name" or the "gene_id" column of the provided GTF files. As the second step, gene pairs with overlapping annotations are identified in each genome. Finally, these gene pairs are removed from all networks (not just from the networks of the species where they were found to overlap!).

For all remaining edges of the networks, the function adds the genomic distance between the 2 genes that form the edge as the edge attribute "genomic_dist". If the genes are annotated on different chromosomes, the distance is set to \code{Inf}. This information can be used for further sanity checking, e.g. to check the relationship between edge weight and genomic proximity.
}
\examples{
network_list_filt <- removeOverlappingGenePairs(network_list_raw,
                                                gtf_list,
                                                clone2species,
                                                "gene_name")
}
